Moved `AuthenticationTestUser` test class to fixture folder.
