+++ tests/Integration/Auth/Fixtures/AuthenticationTestUser.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Illuminate\Tests\Integration\Auth\Fixtures;
+
+use Illuminate\Foundation\Auth\User as Authenticatable;
+
+class AuthenticationTestUser extends Authenticatable
+{
+    public $table = 'users';
+    public $timestamps = false;
+
+    /**
+     * The attributes that are mass assignable.
+     *
+     * @var array
+     */
+    protected $guarded = ['id'];
+
+    /**
+     * The attributes that should be hidden for arrays.
+     *
+     * @var array
+     */
+    protected $hidden = [
+        'password', 'remember_token',
+    ];
+}
